+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## New features
 - Add models for studentLanguageInstructionProgramAssociations, studentHomelessProgramAssociations, and studentTitleIPartAProgramAssociations
 - Add model for Special Education Program Services
+- Add models for cohorts
 - Add model for graduation plans
+
 
 ## Under the hood